third_party/blink/web_tests/platform/generic/http/tests/xmlhttprequest/status-after-abort-expected.txt

Test for rdar://problem/5766352: REGRESSION: XMLHttpRequest.abort() resets response status.

A newly created request. Response status: 0; statusText: ''; readyState: 0
Opening...
Onreadystatechange. Response status: 0; statusText: ''; readyState: 1
Opened request. Response status: 0; statusText: ''; readyState: 1
Sent request. Response status: 0; statusText: ''; readyState: 1
Onreadystatechange. Response status: 200; statusText: 'OK'; readyState: 2
Onreadystatechange. Response status: 200; statusText: 'OK'; readyState: 3
Onreadystatechange. Response status: 200; statusText: 'OK'; readyState: 4
Aborting the request...
After aborting the request. Response status: 0; statusText: ''; readyState: 0
Reopening the request to check that the status is reset...
Onreadystatechange. Response status: 0; statusText: ''; readyState: 1
Done.
